pyproject.toml
[project]
name = "backend"
version = "0.1.0"
dependencies = [
    "django",
    "django-ninja",
    "psycopg[binary]",
    "granian[reload]"
]

[tool.uv.sources]
backend = { path = "." }
